feat: add all logic of home component inside the page, use the components folder only to components
