+++ src/components/Hero/index.tsx
