+++ src/components/index.ts
@@ -15,4 +15,6 @@ export * from '@/components/Layout'
 export * from '@/components/Loading'
 export * from '@/components/Modal'
 export * from '@/components/ProfileName'
+export * from '@/components/Hero'
+export * from '@/components/Articles'
 export * from '@/components/Tabs'
